add del downloadPackage for windows

+++ installDemo/installApplication.py
@@ -190,7 +190,10 @@ def clickMeInstallChrome():  # 当acction被点击时,该函数则生效
             except:
                 tkinter.messagebox.showerror('错误', '安装失败')
                 # installChromeAction.configure(text='Install fail ')
-            os.system("rm -f %s" % packagePWD)
+            if platform.system() == "Windows":
+                os.system("rm -f %s" % packagePWD)
+            else:
+                os.system("del %s" % packagePWD)
         else:
             try:
                 if platform.system() == "Windows":
